Renaming Statement to StatementDecorator, to better express what it is supposed to do.
